fixed insta death bug

+++ include/player.h
@@ -22,7 +22,7 @@ typedef struct Player {
 } Player;
 
 Player *playerCreate(int x, int y, int hp, int max_hp);
-void playerMove(Player *p, size_t x, size_t y);
+void playerMove(Player *p, int x, int y);
 void playerUpdateHP(Player *p, int change);
 void playerUpdatePercentHP(Player *p);
 Rectangle *playerGetRectangle(Player *p);
